% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{coef.slim}
\alias{coef.slim}
\title{Extract Model Coefficients from Singular Linear Model}
\usage{
\method{coef}{slim}(object, ...)
}
\arguments{
\item{object}{an object of class 'slim', usually, a result of a call to
'slim'.}

\item{...}{arguments passed to or from other methods.}
}
\value{
a vector of model coefficients.
}
\description{
Extract Model Coefficients from Singular Linear Model
}
