% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim.R
\name{compute_laurent}
\alias{compute_laurent}
\title{Laurent Expansion of Inverse of Linear Matrix Function}
\usage{
compute_laurent(V, zapsmall = TRUE)
}
\arguments{
\item{V}{for some integer m >= 1, an array of dimension (m, m, 2), where V[,
, 1] is the intercept and V[, , 2] is the slope of the linear matrix
function.}

\item{zapsmall}{logical: should zapsmall be called on the result? Default TRUE.}
}
\value{
array of dimension (m, m, 2), where W[, , 1] corresponds to the
exponent -1, and W[, , 2] corresponds to the exponent 0.
}
\description{
This function computes the first two terms of the Laurent expansion of the
inverse of a linear matrix function.
}
