% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{confint.slim}
\alias{confint.slim}
\title{Confidence Intervals for Model Parameters from Singular Linear Model}
\usage{
\method{confint}{slim}(object, parm, level = 0.95, empirical = TRUE, ...)
}
\arguments{
\item{object}{an object of class 'slim', usually, a result of a call to
'slim'.}

\item{parm}{a specification of which parameters are to be given
    confidence intervals, either a vector of numbers or a vector of
    names.  If missing, all parameters are considered.}

\item{level}{the confidence level required.}

\item{empirical}{logical indicating if empirical variances of y should be
used in estimating standard errors (the default). Empirical standard errors
should be used unless covariances have been well modelled.}

\item{...}{arguments passed to or from other methods.}
}
\value{
A matrix (or vector) with columns giving lower and upper confidence
limits for each parameter.
}
\description{
Confidence Intervals for Model Parameters from Singular Linear Model
}

