% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{dialysis}
\alias{dialysis}
\title{Renal Function in Three Groups of Peritoneal Dialysis Patients}
\format{
A data.table with 116 rows and 5 variables:
\tabular{ll}{
id \tab patient identifier, a character string \cr
group \tab treatment group identifier, a character string \cr
vintage \tab days since starting dialysis, an integer \cr
month \tab month of observation, an integer \cr
renalfn \tab renal function of the patient at that month, numeric \cr
}
}
\source{
This data is derived from the Global Fluid Study. This part of the
study was led by Dr James Chess and Prof. Nick Topley.
}
\usage{
dialysis
}
\description{
Longitudinal data on the renal function of 116 patients observed on up to
five different occasions.
}
\references{
Lambie, M., Chess, J. et al. (2013). Independent effects of
systemic and peritoneal inflammation on peritoneal dialysis survival. J Am
Soc Nephrol, 24, 2071--80.
}
\keyword{datasets}
