% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{fitted.slim}
\alias{fitted.slim}
\title{Extract Model Fitted Values from Singular Linear Model}
\usage{
\method{fitted}{slim}(object, ...)
}
\arguments{
\item{object}{an object of class 'slim', usually, a result of a call to
'slim'.}

\item{...}{arguments passed to or from other methods.}
}
\value{
a vector of fitted values from the model fit.
}
\description{
Extract Model Fitted Values from Singular Linear Model
}

