% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{predict.slim}
\alias{predict.slim}
\title{Model Predictions from Singular Linear Model}
\usage{
\method{predict}{slim}(object, newdata, ...)
}
\arguments{
\item{object}{an object of class 'slim', usually, a result of a call to
'slim'.}

\item{newdata}{An optional data frame in which to look for variables with
    which to predict.  If omitted, the fitted values are used.}

\item{...}{arguments passed to or from other methods.}
}
\value{
a vector of model predictions.
}
\description{
Model Predictions from Singular Linear Model
}
