% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{residuals.slim}
\alias{residuals.slim}
\title{Extract Model Residuals from Singular Linear Model}
\usage{
\method{residuals}{slim}(object, ...)
}
\arguments{
\item{object}{an object of class 'slim', usually, a result of a call to
'slim'.}

\item{...}{arguments passed to or from other methods.}
}
\value{
a vector of model residuals.
}
\description{
Extract Model Residuals from Singular Linear Model
}

