% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim-package.R
\docType{package}
\name{slim-package}
\alias{slim-package}
\title{Singular linear models for longitudinal data.}
\description{
The slim package fits singular linear models to longitudinal data. Singular
linear models are useful when the number, or timing, of longitudinal
observations may be informative about the observations themselves. They are
described in Farewell (2010) <doi:10.1093/biomet/asp068>, and are extensions
of the linear increments model of Diggle et al. (2007)
<doi:10.1111/j.1467-9876.2007.00590.x> to general longitudinal data.
}
\details{
The most important function is slim, whose formula interface is similar to
that of lm.
}
\seealso{
\code{\link{slim}}
}
\author{
\strong{Maintainer}: Daniel Farewell \email{farewelld@cf.ac.uk}

}
\keyword{internal}
