% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{slim.methods}
\alias{slim.methods}
\title{Methods for Singular Linear Model Fits}
\arguments{
\item{object}{an object of class 'slim', usually, a result of a call to
'slim'.}

\item{empirical}{logical indicating if empirical variances of y should be
used in estimating standard errors (the default). Empirical standard errors
should be used unless covariances have been well modelled.}

\item{...}{arguments passed to or from other methods.}
}
\description{
Methods for Singular Linear Model Fits
}
