% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{summary.slim}
\alias{summary.slim}
\title{Summarizing Singular Linear Model Fits}
\usage{
\method{summary}{slim}(object, empirical = TRUE, ...)
}
\arguments{
\item{object}{an object of class 'slim', usually, a result of a call to
'slim'.}

\item{empirical}{logical indicating if empirical variances of y should be
used in estimating standard errors (the default). Empirical standard errors
should be used unless covariances have been well modelled.}

\item{...}{arguments passed to or from other methods.}
}
\value{
an object with class c("slim_summary", "slim") and, in addition to
the usual 'slim' components, coefficient_matrix (the matrix of estimated
coefficients, standard errors, z- and p-values) and empirical (logical
indicating if empirical standard errors have been used)
}
\description{
'summary' method for class 'slim'.
}
