% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{vcov.slim}
\alias{vcov.slim}
\title{Extract Variance-Covariance Matrix from a 'slim' Object}
\usage{
\method{vcov}{slim}(object, empirical = TRUE, ...)
}
\arguments{
\item{object}{an object of class 'slim', usually, a result of a call to
'slim'.}

\item{empirical}{logical indicating if empirical variances of y should be
used in estimating standard errors (the default). Empirical standard errors
should be used unless covariances have been well modelled.}

\item{...}{arguments passed to or from other methods.}
}
\value{
a matrix of the estimated covariances between the parameter
estimates.
}
\description{
'vcov' method for class 'slim'.
}

