% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slim.R
\name{fit_slim}
\alias{fit_slim}
\title{Fitter Function for Singular Linear Models}
\usage{
fit_slim(x, V, y)
}
\arguments{
\item{x}{list of design matrices, one for each subject, all having the same
number of columns.}

\item{V}{list of covariance arrays, one for each subject, matching the
dimensions of y.}

\item{y}{list of response vectors, one for each subject.}
}
\value{
a list with components coefficients (the limiting solution),
residuals, fited_values, vcov_empirical and vcov_modelled.
}
\description{
This function computes the limiting solution to the estimating equation
sum(x' V^{-1} (y - x beta)) = 0 as the covariance V tends from V[, , 1] +
V[, , 2] to V[, , 1].
}

