% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_covariances.R
\name{list_covariances}
\alias{list_covariances}
\alias{list_covariances.character}
\alias{list_covariances.list}
\alias{list_covariances.function}
\alias{list_covariances.jmcmMod}
\alias{list_covariances.lmerMod}
\title{List Covariance Matrices for Every Subject}
\usage{
list_covariances(obj, t)

\method{list_covariances}{character}(obj, t)

\method{list_covariances}{list}(obj, t)

\method{list_covariances}{function}(obj, t)

\method{list_covariances}{jmcmMod}(obj, t)

\method{list_covariances}{lmerMod}(obj, t)
}
\arguments{
\item{obj}{an R object of class character, function, or a model fit}

\item{t}{list of vectors of observation times, one for each subject}
}
\value{
a list containing covariance matrices of appropriate dimensions
}
\description{
This function is generic, and methods exists for character, list, function,
and various model fit classes.
}
